modified ex07.c: var int to float

+++ lista3/ex07.c
@@ -2,15 +2,15 @@
 
 int main(void) {
     
-    int maior, menor;
+    float maior, menor;
 
     maior = 21;
     menor = 21;
 
     if(maior > menor) {
-        printf("maior: %i\n", maior);
+        printf("maior: %f\n", maior);
     }else if (maior < menor) {
-        printf("maior: %i\n", menor);
+        printf("maior: %f\n", menor);
     } else {
         printf("Valores iguais\n");
     }
